IrcClientConfig does not know about encryption
Passwords are kept in config objects unencrypted; passwords are encrypted when the config itself
is stored in the database and dencrypted when the config is retrieved from the database.
