+++ config.sample.yaml
@@ -329,13 +329,12 @@ ircService:
     # Default: 300 seconds (5 mins)
     requestTimeoutSeconds: 300
 
-  # WARNING - STORING PASSWORDS ENCRYPTED IS PRECAUTIONARY TO MAKE COMPROMISE LESS
-  #           LIKELY. THIS MEASURE IS ONLY AS SECURE AS THE KEY FILE, BELOW.
+  # WARNING - STORED PASSWORDS ENCRYPTED ARE COMPLETELY RETRIEVABLE WITH THE KEY.
+  #           THE BRIDGE NEEDS TO SEND PASSWORDS TO THE IRC SERVERS AS PLAINTEXTS.
   #
-  # The path to the RSA PEM-formatted key to use when encrypting and decrypting
-  # IRC passwords that have been stored in the database. Passwords are stored by
-  # using the admin room `!storepass server.name passw0rd. When a connection is
-  # made for the corresponding virtual IRC user of the matrix user that has issued
-  # !storepass, they will connect with the password that has been persisted, if
-  # there is one.
+  # The path to the RSA PEM-formatted private key to use when encrypting IRC passwords
+  # for storage in the database. Passwords are stored by using the admin room command
+  # `!storepass server.name passw0rd. When a connection is made for the corresponding
+  # virtual IRC user of the matrix user that has issued !storepass, they will connect
+  # with the (decrypted) password that has been persisted, if there is one.
   passwordEncryptionKeyPath: "passkey.pem"
