+++ lib/DataStore.js
@@ -2,6 +2,7 @@
 "use strict";
 
 var Promise = require("bluebird");
+var crypto = require('crypto');
 
 var MatrixRoom = require("matrix-appservice-bridge").MatrixRoom;
 var MatrixUser = require("matrix-appservice-bridge").MatrixUser;
@@ -52,6 +53,8 @@ function DataStore(userStore, roomStore, pkeyPath) {
         unique: true,
         sparse: false
     }, errLog("user id"));
+
+    this._privateKey = null;
 
     if (pkeyPath) {
         try {
@@ -444,7 +447,22 @@ DataStore.prototype.getIrcClientConfig = function(userId, domain) {
         if (!configData) {
             return null;
         }
-        return new IrcClientConfig(userId, domain, configData);
+        let clientConfig = new IrcClientConfig(userId, domain, configData);
+
+        if (clientConfig.getPassword()) {
+            if (!this._privateKey) {
+                throw new Error(`Cannot decrypt password of ${userId} - no private key`);
+            }
+
+            let decryptedPass = crypto.privateDecrypt(
+                this._privateKey,
+                new Buffer(this._config.password, 'base64')
+            ).toString();
+
+            clientConfig.setPassword(decryptedPass);
+        }
+
+        return clientConfig;
     });
 };
 
@@ -455,6 +473,23 @@ DataStore.prototype.storeIrcClientConfig = function(config) {
             user = new MatrixUser(config.getUserId());
         }
         var userConfig = user.get("client_config") || {};
+
+        if (userConfig.getPassword()) {
+            if (!this._privateKey) {
+                throw new Error(
+                    'Cannot store plaintext passwords'
+                );
+            }
+
+            let encryptedPass = crypto.publicEncrypt(
+                this._privateKey,
+                new Buffer(password)
+            ).toString('base64');
+
+            // Store the encrypted password, ready for the db
+            userConfig.setPassword(encryptedPass);
+        }
+
         userConfig[config.getDomain().replace(/\./g, "_")] = config.serialize();
         user.set("client_config", userConfig);
         return this._userStore.setMatrixUser(user);
@@ -462,18 +497,9 @@ DataStore.prototype.storeIrcClientConfig = function(config) {
 };
 
 DataStore.prototype.storePass = Promise.coroutine(
-    function*(userId, domain, pass, forcePlaintextStorage) {
+    function*(userId, domain, pass) {
         let config = yield this.getIrcClientConfig(userId, domain);
-        if (!this._privateKey && !forcePlaintextStorage) {
-            throw new Error(
-                'WARNING: Store plaintext passwords at your own ' +
-                'risk; use -f to force'
-            );
-        }
 
-        if (this._privateKey && !config._privateKey) {
-            config.setPrivateKey(this._privateKey);
-        }
         config.setPassword(pass);
 
         yield this.storeIrcClientConfig(config);
