+++ lib/bridge/IrcBridge.js
@@ -138,16 +138,6 @@ IrcBridge.prototype.createBridgedClient = function(ircClientConfig, matrixUser, 
         );
     }
 
-    if (ircClientConfig._config.isPasswordEncrypted) {
-        if (!this.getStore()._privateKey) {
-            throw new Error(
-                `Cannot create BridgeClient with encrypted password unless a private key ` +
-                `is provided to the bridge. See config 'passwordEncryptionKeyPath'.`
-            );
-        }
-        ircClientConfig.setPrivateKey(this.getStore()._privateKey);
-    }
-
     return new BridgedClient(
         server, ircClientConfig, matrixUser, isBot,
         this._ircEventBroker, this._identGenerator, this._ipv6Generator
@@ -172,9 +162,7 @@ IrcBridge.prototype.run = Promise.coroutine(function*(port) {
     let pkeyPath = this.config.ircService.passwordEncryptionKeyPath;
 
     this._dataStore = new DataStore(
-        this._bridge.getUserStore(),
-        this._bridge.getRoomStore(),
-        pkeyPath
+        this._bridge.getUserStore(), this._bridge.getRoomStore(), pkeyPath
     );
     yield this._dataStore.removeConfigMappings();
     this._identGenerator = new IdentGenerator(this._dataStore);
