+++ lib/bridge/MatrixHandler.js
@@ -457,8 +457,7 @@ MatrixHandler.prototype._onAdminMessage = Promise.coroutine(function*(req, event
         let notice;
 
         try {
-            let force = args[1] === '-f';
-            yield this.ircBridge.getStore().storePass(userId, domain, args[0], force);
+            yield this.ircBridge.getStore().storePass(userId, domain, args[0]);
             notice = new MatrixAction(
                 "notice", `Successfully stored password`
             );
