+++ lib/models/IrcClientConfig.js
@@ -1,7 +1,5 @@
 "use strict";
 
-const crypto = require('crypto');
-
 /**
  * Contains IRC client configuration, mostly set by Matrix users. Used to configure
  * IrcUsers.
@@ -56,34 +54,11 @@ class IrcClientConfig {
     }
 
     setPassword(password) {
-        if (!this._privateKey) {
-            this._config.password = password
-            this._config.isPasswordEncrypted = false;
-            return;
-        }
-
-        let encryptedPass = crypto.publicEncrypt(
-            this._privateKey,
-            new Buffer(password)
-        );
-
-        this._config.password = encryptedPass.toString('base64');
-        this._config.isPasswordEncrypted = true;
+        this._config.password = password;
     }
 
     getPassword() {
-        if (!this._config.isPasswordEncrypted) {
-            return this._config.password;
-        }
-
-        if (!this._privateKey) {
-            throw new Error(`Cannot decrypt password of ${this.userId} - no private key given`);
-        }
-        console.info('Decrypting password with pkey' + this._privateKey.toString());
-        return crypto.privateDecrypt(
-            this._privateKey,
-            new Buffer(this._config.password, 'base64')
-        ).toString();
+       return this._config.password;
     }
 
     setDesiredNick(nick) {
